tableview pagination and searchFieldsEditor added administration routing, inline css and preference service fixed layout adjustments

+++ org.eclipse.mdm.application/src/main/webapp/src/app/administration/preference.component.ts
@@ -46,7 +46,11 @@ export class PreferenceComponent implements OnInit, OnDestroy {
   }
 
   removePreference( preference: Preference ) {
-      this.preferenceService.deletePreference( preference.id );
+      if (preference.id) {
+        console.log(preference.id);
+        this.preferenceService.deletePreference( preference.id )
+        .subscribe();
+      }
       let index = this.preferences.findIndex( p => p === preference );
       this.preferences.splice( index, 1 );
   }
